Parsing out git context client-side for build Moved some code to the top of the subcommand trait. They are be reused
