+++ cli/orb/src/subcommand/Cargo.toml
@@ -14,6 +14,7 @@ orbital_headers = { path = "../../../../models/orbital_headers" }
 orbital_services = { path = "../../../../services" }
 log = "0.4"
 tonic = { version = "0.1.0-alpha.6", features = [ "rustls"] }
+git2 = "0.10"
 
 [lib]
 name = "subcommand"
