+++ models/protos/build_meta.proto
@@ -11,7 +11,9 @@ message BuildTarget {
     string remote_uri = 4;
     string branch = 5;
     string commit_hash = 6;
-    int32 id = 7;
+    string docker_image = 7;
+    string envs = 8;
+    int32 id = 9;
 }
 
 message BuildStage {
